move configs gamechangerml/ -> gamechangerml/src - to comply w/ development rules in readme
