+++ gamechangerml/api/getInitModels.py
@@ -9,7 +9,7 @@ from os.path import join, exists
 from gamechangerml.api.fastapi.model_config import Config
 from gamechangerml.src.data_transfer import get_latest_model_name
 from gamechangerml.src.services import S3Service
-from gamechangerml.configs import S3Config
+from gamechangerml.src.configs import S3Config
 from gamechangerml.src.utilities import configure_logger
