+++ gamechangerml/scripts/data_transfer/download_eval_data.py
@@ -1,7 +1,7 @@
 """Script to download evaluation data. Requires command-line input."""
 
 from gamechangerml.src.services import S3Service
-from gamechangerml.configs import S3Config
+from gamechangerml.src.configs import S3Config
 from gamechangerml.src.data_transfer import download_eval_data
 from gamechangerml.src.utilities import configure_logger
