+++ gamechangerml/src/model_testing/validation_data.py
@@ -3,7 +3,7 @@ import numpy as np
 from collections import OrderedDict
 from gamechangerml.src.utilities.text_utils import normalize_answer, normalize_query, filter_title_queries
 from gamechangerml.src.utilities.test_utils import *
-from gamechangerml.configs import ValidationConfig
+from gamechangerml.src.configs import ValidationConfig
 from gamechangerml.api.utils.logger import logger
 from gamechangerml.src.utilities.test_utils import filter_date_range, open_txt
 from gamechangerml import REPO_PATH
